Overhaul configurations
Get rid of the notion of a single `Default-Configuration`, and instead
have a priority order list of configurations. This is important as we
can't know what configuration(s) will be installed. Also, remove the
notion of a `Default-Configuration` attribute value being used even if
some other component of a configuration is being used, as this is likely
to be brittle. Tools instead shall pick a configuration and use those
attributes (falling back to common attributes).