Allow specification of required components
Add "Components" attribute applying to a <requirement> object, to allow
a CPS to specify what components of a transitive dependency are
required. This is mainly useful for improved diagnostics (and also
to serve as a hint that consumers should likewise be able to state an
up-front requirement on specific components of a package).