Tweaks to merging and search
Consolidate search paths for improved consistency, and add optional
version subdirectory for all platforms. Add ability to separate a CPS
into multiple non-configuration-specific files in order to allow
installation of component subsets. Specify a way to find CPS files if
version information is included in the file name (which will help with
co-location even when a version subdirectory is not used).