Add "non-normative" extras
Add an initial list of compiler features; while not strictly normative,
it is strongly recommended that implementations support these in a
consistent manner. Add a list of supplemental attributes which may be
useful but are not required to be supported. (Both of these have been
planned, but weren't written until now.)