Add a bunch of platform stuff
Add platform object and related attributes. Add a note on platform
matching. Add a "recommendations" section explaining how to accomplish
various common (or not so common) tasks using CPS, particularly several
that deal with platform quirks. (This will also serve as a guide to
problems that we want to address via usage patterns rather than schema.)