version schema "semver" should mean "semver"
Currently "Semver" is used to mean x.y.z, but with not that a project's
versioning should use semantic versioning. It then includes that a
project should manually set the oldest comaptible version. This is
exactly the sort of thing semver is supposed to fix.

Therefore, this patch makes two changes:
1) semver is changed to actually mean semver. When it is set
   Compat-Version should not be set, and the versioning should be
   determined using semantic versioning rules.
2) A new format, "x.y.z" is introduced as the new default. It allows
   sem-version like ordering, but without following semver. This is
   the currentlys speced behavior for "semver"