Add 16×16 logo, favicon
Clean up logo SVGs (this is just tidying the markup; no visual changes).
Add a 16×16 version of the logo, which is the same as the existing logos
but has been tweaked to help it render better at such a small size. Use
this to (finally!) generate a favicon for the HTML.