Adjust how searching works
Allow `-` as a separator in e.g. `<name>-<component>.cps`. This is
required as `:` is not available on e.g. Windows. (We might be able to
use some other character, but even `-` introduces a disjoint convention,
as `:` is otherwise used as the component separator. However, there is
at least precedent for using `-` as a separator.)

In order to avoid serious confusion in light of this change, remove
support for `<name>-*.cps`. This also means that the CPS file name
relative to the package name is much more deterministic. In order to
continue to support multiple versions of a package, add additional
search paths and generally clean up the set of search paths to be more
consistent.

Arguably, this should entail a major version bump, although as we are
pre-1.0, we are probably going to make explicit at some point that,
prior to 1.0, breaking changes may occur at minor version bumps. (In any
case, things are likely to remain "rough" while we are in the early
stages of roll-out.)