Wrap "code"
Adjust CSS to allow "code" to be hyphenated and to wrap at hyphens or
whitespace. This avoids poorly formatted text in case of a long run of
"code". Note that "code" here means highlighted entities such as types,
keywords, and especially object type names, where the latter are
especially likely to be long enough to warrant being split.