Use SPDX expressions for license and default_license
Currently, SPDX license identifiers are used, but a custom composition
syntax of nested lists is used. This patch replaces that custom syntax
with full SPDX expressions. SPDX has become an ISO standard since work
on CPS began, making it a good candidate for use in CPS.

For example, the expression in the documentation using the nested list
syntax `["CC-BY-4.0", ["GPL-2.0", "LGPL-3.0-or-later"]]` would be
written as: `CC-BY-4.0 AND (GPL-2.0-only OR LGPL-3.0-or-later)`.