searching: when searching by name specify that given case then lower case
In the event one searches for `MyPackage`, and there is a
`MyPackage.cps` and a `mypackage.cps`, the specification should ensure a
stable sort order. In this case, since `MyPackage` was searched for, an
implementation should use `MyPackage.cps` before `mypackage.cps`

The change from "proper case" to "given case" is intention, as it may
not be obvious what the proper case is. For example, "LLVM", "OpenGL",
"libxml2",  "Python" are all proper capitalizations of various projects.