Remove some name-first search paths
Remove search paths that place the package name before any path
component indicating that CPS files are to be found within. This removes
most of the paths from 0b8f2ffb1c13. This may affect some existing
projects which, at least for CMake files, are using the equivalent of
the paths being removed, but we can revisit if necessary. Meanwhile,
this improves consistency and reduces the load on tools trying to find
CPS files.