Implement automatic HTML generation
Add automation via GitHub Actions to generate the specification HTML and
deploy the same to GitHub Pages.

Additionally, rework HTML generation to use Poetry¹ to set up and manage
a virtual environment used to generate the HTML. This helps ensure that
both local and automated builds are using a uniform environment.

Note that deployment eschews the `upload-pages-artifact`² action because
the mechanism used here allows us to bundle the files and correct their
permissions in one command.

For clarity, the `_ext` directory is renamed to `_extensions`. Also, the
minimum Sphinx version is bumped to 6.2, as that's what's been in use
recently and Poetry allows us to be less "stuck" on what's provided by
distributions.

¹ https://python-poetry.org/
² https://github.com/actions/upload-pages-artifact/

Fixes #40.

Co-authored-by: Matthew Woehlke <mwoehlke.floss@gmail.com>