Make compile definitions key/value
Rewrite the specification of `definitions` to separate the definition
name and value. In order to keep things sane for parsers, this
necessitates always specifying definitions as a language-specific map.
Also add an explicit 'all languages' key (which we probably should have
had all along).

In theory, this makes it easier for consumers to perform  certain
operations such as querying whether a particular symbol is defined. It
should also slightly simplify combining definitions from multiple
components, or language-specific definitions with non-language-specific
definitions (although, in practice, redefinitions on the command line
are likely to be harmless).

Note that one potential complication of this approach is that it
requires null values to be handled specially. This is required because
we need a way to distinguish between `-Dfoo` and `-Dfoo=`, which may
have different effects.

At least for now, the ability to specify explicitly undefined symbols
has been dropped. Besides being rare, and potentially problematic for
users, it turns out that tools may not support these either. (At least
in theory, `compile_flags` can be used if necessary.)