Add link to JSON schema
Add a link to the generated JSON schema from the main documentation.

This required a bit of fiddling to be able to link to a file that is
generated, but not by Sphinx (proper), as there does not seem to be any
built-in way to give a relative link to something that Sphinx doesn't
know about. On the plus side, having to concoct such a mechanism allowed
providing a role that obtains the proper file name from the site
configuration.

This is a very odd role, in that the input is taken as the link text,
and the link reference is controlled by the site configuration, but it
works for now. A better solution might be to provide a built-in target
definition, but I'm not sure how to do that, or if it would permit
"undefined" local links. (The approach here sort of "tricks" Sphinx into
thinking the link is external.)