Add attribute defaults, format
Make default values part of the "machine readable" portion of an
attribute specification. This makes them stand out more in the
documentation, and also allows them to be exported to the JSON schema.
Also add value format, which is exported to the JSON schema but does not
appear in the main documentation.