Fix type of "definitions" attribute
Adjust the specified type of the "definitions" attribute to reflect that
null values are allowed (as specified in the prose). Improve type
parsing to be more robust about how union types are split. Add support
for 'nullable' types.

Currently, we don't really support prose generation for compound types
where the inner value is a union type. (Schema generation doesn't have
this limitation and can handle these just fine.) The 'nullable' type is
a bit of a hack for this specific instance, but for now it's good enough
for the types that actually exist in the schema.