search: clarify that the case of the `name` is as requested or all lower
This means that if the requester (whether that is directly from a user
or from another CPS file) specifies the name in a case that is not all
lower, search for that exactly as well as the all lower.

For example, if "OpenGL" is requested, then the name should be searched
as "OpenGL" and as "opengl".