Update/expand CPS history
As part of some recent(ish) external publications, various parties made
another attempt at describing the history of CPS. Those writings ended
up significantly pared down, but since the information is useful, it was
decided to incorporate it into our "official" history in the CPS repo
itself.

Use that information, mainly authored by Vito Gamberini (with many
thanks for the same), to update and expand our existing entry. This adds
a more detail (and a few minor corrections with respect to timing of
pre-CPS events).

Co-Authored-By: Vito Gamberini <vito.gamberini@kitware.com>