Require "prefix", allow absolute prefixes
Add a semi-optional `prefix` attribute which can be used to specify an
absolute prefix (for non-relocatable packages). Require that exactly one
of `cps_path` or the new `prefix` is supplied. Improve documentation on
how to determine the prefix from `cps_path`, and drop no-longer-needed
guessing mechanisms. Remove "install" from "package [install] prefix",
as packages aren't necessarily "installed".

This is slightly more complicated in some ways, but does enforce that
the CPS file itself always provides a substitution for `@prefix@`, which
allows for some other simplifications. Note that, despite the seemingly
minimal change to the schema, this is in some respects a significant
breaking change in that something formerly optional is now required.