Expand "style" advice, make own section
Create a new "Source Presentation" section under "Contributing", which
expands slightly on the advice for how to properly "style" raw markup.
In particular, add a reference to sembr.org, which is a semi-formalized
presentation of similar guidelines. (I actually believe I came up with
the system I use independently, but it has rather obvious benefits for
version controlled prose, and being able to point out that other people
are using similar systems is always beneficial.)

Closes #87.