configurations: Remove support for `:` as a separator
`:` noted to be invalid on some platforms, notably on Windows. As such,
one of the two biggest platform targets for CPS will not support `:`,
and tools targeting both Windows and Linux or MacOS will likely not
write the files, given that they would need two separate code paths to
do so. Additionally, having a character that is known to not work on
some platforms complicates cross compilation with binaries. For example,
if a dependency is generated on Linux, then pulled onto a Windows system
to do a Windows -> Linux cross compile, A CPS file that was valid on
Linux containing a `:` will not work.

Since there is already an alternative (using `-`), the most obvious
solution is to simply drop `:` and only use `-`.